@@ -1,5 +1,7 @@
 <?php
 
-require '../vendor/autoload.php';
+require_once __DIR__.'/../vendor/autoload.php';
 
-echo (new Kirby())->render();
+$kirby = new Kirby;
+
+echo $kirby->render();
