@@ -2,9 +2,10 @@
 
 return [
     'debug' => true,
-    'bnomei.securityheaders.enabled' => 'force', // force even on http localhost
+
+    'bnomei.securityheaders.enabled' => true, // force even on http localhost
 
     'bnomei.securityheaders.loader' => function () {
-        return kirby()->roots()->site() . '/loader-test.json';
+        return kirby()->roots()->config().'/csp.json';
     },
 ];
