@@ -1,3 +1,3 @@
 <?php
 
-@include_once __DIR__ . '/../../../../index.php';
+@include_once __DIR__.'/../../../../index.php';
